requirements.txt
RPi.GPIO==0.6.5
influxdb==5.2.1
